+++ proxy_test.go
@@ -367,7 +367,7 @@ type fakeObjectStore struct {
 	attributesFunc     func(context.Context, string) (objectMetadata, error)
 	newReaderFunc      func(context.Context, string) (objectRead, error)
 	newRangeReaderFunc func(context.Context, string, int64, int64) (objectRead, error)
-	newWriterFunc      func(context.Context, string, objectWriteOptions) io.WriteCloser
+	newWriterFunc      func(context.Context, string, objectWriteOptions) objectWriter
 }
 
 func (s *fakeObjectStore) attributes(
@@ -408,7 +408,7 @@ func (s *fakeObjectStore) newWriter(
 	ctx context.Context,
 	objectName string,
 	options objectWriteOptions,
-) io.WriteCloser {
+) objectWriter {
 	if s.newWriterFunc == nil {
 		panic("unexpected newWriter call")
 	}